% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode -single latitude.R
\name{Geocode_single}
\alias{Geocode_single}
\title{Function for Hurricane visualization with wind speed span at particular latitude.}
\usage{
Geocode_single(tidy_tracks, stormvalue, cname, latitude1)
}
\arguments{
\item{-}{data frame name which should have following columns}
}
\value{
- graphical representation
}
\description{
NAME
}
\details{
: Data (inggggg form of data frame) fed to this function is enriched and data visulaization layer uses enriched data to plot the parameters.
}
\note{
More descriptive
}
\examples{
- Geocode_single(tidy_tracks)
}
